Scripts/Misc: Migrate world event scripts to separate files 2
